First revamp done with the following changes:
`App.jsx`

1. Removed redundant state

   - The `data` state formerly used to populate the chart was a [redundant state](https://react.dev/learn/choosing-the-state-structure#avoid-redundant-state). The (chart) data can simply be computed during rendering.

2. Removed unneeded effect
   - If an operation - in this case, updating the local storage - can be 'tied' to a user event - in this case, submitting a new entry - then carry it out in the handler of that event; [`useEffect` should not be the first go-to](https://react.dev/learn/you-might-not-need-an-effect#how-to-remove-unnecessary-effects).

   - [You don't need Effects to transform data for rendering](https://react.dev/learn/you-might-not-need-an-effect#how-to-remove-unnecessary-effects). The old way of transforming the chart data in an effect was replaced with the `getChartData` function which does the transform during rendering.

   - The old way - i.e. `setData(newData)` in `useEffect` - caused double rendering for the `App` component on app load and for every new entry submitted because the [Effect immediately updates the state](https://react.dev/learn/you-might-not-need-an-effect#how-to-remove-unnecessary-effects) (i.e. `data`): the double rendering on app load can be confirmed from [the newly-added test in the `num-of-renders-test` branch](https://github.com/GermaVinsmoke/bmi-calculator/compare/master...lanresmith:bmi-calculator:num-of-renders-test?expand=1) which fails.

3. 'Undo' was handled in a better and history-like manner - continuing to click `Undo` continues to roll back deletions successively; from the most recent to the least recent. Also, no (unnecessary) display of the `Undo` button on app load e.g. in the case of immediately refreshing the app after a deletion - as every app load should be seen like a new session.

`Bar.jsx`

1. `labelData` and `bmiData` were consolidated into `chartData`, for compactness and a clearer variable name (hence, clearer purpose). Similar to ["Group related state"](https://react.dev/learn/choosing-the-state-structure#group-related-state) - yeah, these ones are props though, there's no harm however in following the principle regarding them too :).

`BmiForm.jsx`

1. Form was used so that the enter key could simply be hit for faster data entry and more accessibility - whatever is submitted as a form should be wrapped in `<form>` tags.

Necessary updates to existing tests were carried out and new tests were added.

**PS**:
I am aware that this could have been done for persisting to local storage:

```
useEffect(() => {
   storeData('data', state);
}, [state]);
```

rather than the current way of handling the persistence to local storage which is quite verbose: calling `storeData` in each of the handlers of the three events - new submission, delete and undo.

The Effect way of handling the persistence to local storage however runs once on app load, even before any user event, essentially storing back to the local storage _exactly_ what we just retrieved from it (with `getData('data')`). While this is not really a problem - in this case - this is something that we don't really intend.

We would like to be more intentional with the results of our actions - again, _if an operation can be tied to a user event, then carry it out in the handler of that event_. This way, we're sure of what is causing what. [By the time an Effect runs, you don't know what the user did](https://react.dev/learn/you-might-not-need-an-effect#how-to-remove-unnecessary-effects).